+++ src/searchDotCom.tsx
@@ -3,6 +3,6 @@ import { LaunchProps } from "@raycast/api";
 import DotComCommand from "./components/DotComCommand";
 import SearchCommand from "./components/SearchCommand";
 
-export default function SearchCloud(props: LaunchProps) {
+export default function SearchDotCom(props: LaunchProps) {
   return <DotComCommand Command={SearchCommand} props={props} />;
 }