+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **shortcuts**: The `Cmd-Enter` shortcut used for options like "Open Query in Browser" is now `Cmd-Shift-Enter` to avoid conflicts with reserved Raycast shortcuts.
 - **all**: Proxying requests over Unix domain socket for custom Sourcegraph instances is now supported via the new `Sourcegraph Instance: Proxy` option; note that HTTP proxies are not yet supported ([#21](https://github.com/bobheadxi/raycast-sourcegraph/pull/21))
 - **notebooks**: Notebooks have been disabled on Sourcegraph.com for some time now, so the command for interacting with public notebooks on Sourcegraph.com has been removed. The "Find Search Notebooks" command for custom instances is still available for the time being.
+- **internal**: The extension now exports some basic telemetry to the connected Sourcegraph instance only. If you are using a self-hosted or on-prem instance, telemetry will be managed according to the instance's telemetry settings. This can be disabled in the extension preferences.
 - **internal**: Updated dependencies.
 
 ## [Code Search History and improved search results metadata](https://github.com/raycast/extensions/pull/4018) - 2022-12-23